+++ vim-engine/src/main/kotlin/com/maddyhome/idea/vim/options/Option.kt
@@ -9,17 +9,14 @@
 package com.maddyhome.idea.vim.options
 
 import com.maddyhome.idea.vim.ex.ExException
-import com.maddyhome.idea.vim.option.NumberOption
 import com.maddyhome.idea.vim.vimscript.model.datatypes.VimDataType
 import com.maddyhome.idea.vim.vimscript.model.datatypes.VimInt
 import com.maddyhome.idea.vim.vimscript.model.datatypes.VimString
+import com.maddyhome.idea.vim.vimscript.model.datatypes.parseNumber
 import java.util.*
 
-/**
- * COMPATIBILITY-LAYER: switched from sealed to abstract
- * Please see: https://jb.gg/zo8n0r
- */
-/*sealed*/abstract class Option<T : VimDataType>(val name: String, val abbrev: String, private val defaultValue: T) {
+// Note that we don't want a sealed hierarchy, so we can add options with custom validation
+abstract class Option<T : VimDataType>(val name: String, val abbrev: String, private val defaultValue: T) {
 
   open fun getDefaultValue(): T {
     return defaultValue
@@ -49,14 +46,6 @@ import java.util.*
       }
     }
   }
-
-  /**
-   * COMPATIBILITY-LAYER: Method added
-   * Please see: https://jb.gg/zo8n0r
-   */
-  @Suppress("PLATFORM_CLASS_MAPPED_TO_KOTLIN")
-  open val value: java.lang.Boolean
-    get() = TODO()
 
   // todo 1.9 should return Result with exceptions
   abstract fun checkIfValueValid(value: VimDataType, token: String)
@@ -128,6 +117,32 @@ open class StringOption(name: String, abbrev: String, defaultValue: VimString, p
   }
 }
 
+open class NumberOption(name: String, abbrev: String, defaultValue: VimInt) :
+  Option<VimInt>(name, abbrev, defaultValue) {
+  constructor(name: String, abbrev: String, defaultValue: Int) : this(name, abbrev, VimInt(defaultValue))
+
+  override fun checkIfValueValid(value: VimDataType, token: String) {
+    if (value !is VimInt) {
+      throw ExException("E521: Number required after =: $token")
+    }
+  }
+
+  override fun getValueIfAppend(currentValue: VimDataType, value: String, token: String): VimInt {
+    val valueToAdd = parseNumber(token) ?: throw ExException("E521: Number required after =: $token")
+    return VimInt((currentValue as VimInt).value + valueToAdd)
+  }
+
+  override fun getValueIfPrepend(currentValue: VimDataType, value: String, token: String): VimInt {
+    val valueToAdd = parseNumber(token) ?: throw ExException("E521: Number required after =: $token")
+    return VimInt((currentValue as VimInt).value * valueToAdd)
+  }
+
+  override fun getValueIfRemove(currentValue: VimDataType, value: String, token: String): VimInt {
+    val valueToAdd = parseNumber(token) ?: throw ExException("E521: Number required after =: $token")
+    return VimInt((currentValue as VimInt).value - valueToAdd)
+  }
+}
+
 open class UnsignedNumberOption(name: String, abbrev: String, defaultValue: VimInt) :
   NumberOption(name, abbrev, defaultValue) {
 
@@ -138,5 +153,27 @@ open class UnsignedNumberOption(name: String, abbrev: String, defaultValue: VimI
     if ((value as VimInt).value < 0) {
       throw ExException("E487: Argument must be positive: $token")
     }
+  }
+}
+
+class ToggleOption(name: String, abbrev: String, defaultValue: VimInt) : Option<VimInt>(name, abbrev, defaultValue) {
+  constructor(name: String, abbrev: String, defaultValue: Boolean) : this(name, abbrev, if (defaultValue) VimInt.ONE else VimInt.ZERO)
+
+  override fun checkIfValueValid(value: VimDataType, token: String) {
+    if (value !is VimInt) {
+      throw ExException("E474: Invalid argument: $token")
+    }
+  }
+
+  override fun getValueIfAppend(currentValue: VimDataType, value: String, token: String): VimInt {
+    throw ExException("E474: Invalid argument: $token")
+  }
+
+  override fun getValueIfPrepend(currentValue: VimDataType, value: String, token: String): VimInt {
+    throw ExException("E474: Invalid argument: $token")
+  }
+
+  override fun getValueIfRemove(currentValue: VimDataType, value: String, token: String): VimInt {
+    throw ExException("E474: Invalid argument: $token")
   }
 }
